requirements.txt
Django==4.2.14
djangorestframework==3.15.2
psycopg2-binary==2.9.9
bcrypt==4.1.3
PyJWT==2.9.0
python-dotenv==25.3